RavenDB-19198 Split about view and license limit to components
