+++ src/Raven.Studio/typescript/viewmodels/database/tasks/EditSubscriptionTaskInfoHub.tsx
@@ -75,7 +75,7 @@ export function EditSubscriptionTaskInfoHub() {
     );
 }
 
-export const defaultFeatureAvailability: FeatureAvailabilityData[] = [
+const defaultFeatureAvailability: FeatureAvailabilityData[] = [
     {
         featureName: "Concurrent Data Subscriptions",
         featureIcon: "data-subscriptions",
